finished price and mileage calculation

+++ SuperFrogSchedule/api/models.py
@@ -167,7 +167,7 @@ class Appearance(Event):
     description = models.CharField(max_length = 1000, blank = True)
     status = models.CharField(max_length = 255, default = "Pending")
     customer = models.ForeignKey(Customer, on_delete = models.SET_NULL, null=True, blank=True)
-    mileage = models.IntegerField(default = 0)
+    mileage = models.DecimalField(default = 0.00, decimal_places=2, max_digits=10)
     cost = models.DecimalField(default = 0.00, decimal_places=2, max_digits=10)
     receipt_number = models.CharField(max_length = 255, blank = True, null=True)
     compensation_date = models.DateTimeField(blank=True, null=True)
